requirements.txt
selenium~=4.31.0
pytest~=8.3.5
webdriver-manager~=4.0.2
dotenv~=0.9.9
python-dotenv~=1.1.0
python-telegram-bot~=22.0
requests~=2.32.3